Add config file parser
Added code to ConfigParser to read in data from a specified file and
populate the various objects in the system.

System data file should be a comma separate list of the form:
<object name>,<mass>,<semi-major axis>[,<eccentricity>]

also adds a sample config file and adds code to read and print the data
in the main() function.